requirements.txt
urllib3==1.26.14
Pillow==9.4.0
keras==2.10.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Flask==2.2.2
Flask-Cors==3.0.10
tensorflow==2.10.0
uuid==0.2.7
plot-keras-history==1.1.38
opencv-contrib-python==3.4.2.16
opencv-python==3.4.2.16
seaborn==0.12.2
matplotlib==3.5.3
matplotlib-inline==0.1.6
pandas==1.1.5
numpy==1.21.6
scikeras==0.10.0
scikit-fuzzy==0.4.2
scikit-image==0.16.2
scikit-learn==0.22
scipy==1.7.3
h5py==3.7.0